services/user-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
werkzeug==3.1.3
python-jose[cryptography]==3.3.0
